add options

+++ fc-solve/scripts/plot-benchmarks-against-time--gen-charts.bash
@@ -1,32 +1,58 @@
 #!/usr/bin/env bash
 set -e -x
-SSCONV_TEST_USE="${SSCONV_TEST_USE:-$1}"
+force="${SSCONV_TEST_FORCE:-false}"
+while test $# -gt 0; do
+	flag="$1"
+	shift
+	arg="$1"
+	shift
+	case "$flag" in
+	--viewer)
+		SSCONV_TEST_VIEWER="$arg"
+		;;
+	--input-format)
+		SSCONV_TEST_USE="$arg"
+		;;
+	--force)
+		force="$arg"
+		;;
+	*)
+		echo "Unknown flag <<$flag>>"
+		exit 1
+		;;
+	esac
+done
 base="Freecell-Solver--benchmarks-vs-time"
 tsv="$base.tsv"
-perl fc-solve/scripts/plot-benchmarks-against-time.pl | tee "$tsv"
+if $force || test \! -e "$tsv"; then
+	perl fc-solve/scripts/plot-benchmarks-against-time.pl | tee "$tsv"
+fi
 xlsx="$base.xlsx"
-cat "$tsv" | (perl -nalE 'say join",",@F[0,2]') | csv2chart xlsx -o "$xlsx" --title "fc-solve Benchmark" --width 1200 --height 600
+if $force || test \! -e "$xlsx"; then
+	cat "$tsv" | (perl -nalE 'say join",",@F[0,2]') | csv2chart xlsx -o "$xlsx" --title "fc-solve Benchmark" --width 1200 --height 600
+fi
 ods="$base.ods"
-ssconvert "$xlsx" "$ods"
+if $force || test \! -e "$ols"; then
+	ssconvert "$xlsx" "$ods"
+fi
 gnum="$base.gnumeric"
-ssconvert "$xlsx" "$gnum"
+if $force || test \! -e "$gnum"; then
+	ssconvert "$xlsx" "$gnum"
+fi
 svgtemp="i.svg"
 svgtemp0="$svgtemp.0"
-if test -z "$SSCONV_TEST_USE"
-then
-    ssconvert --export-graphs "$gnum" "$svgtemp"
-elif test "$SSCONV_TEST_USE" = xlsx
-then
-    ssconvert --export-graphs "$xlsx" "$svgtemp"
-elif test "$SSCONV_TEST_USE" = ods
-then
-    ssconvert --export-graphs "$ods" "$svgtemp"
+if test -z "$SSCONV_TEST_USE"; then
+	ssconvert --export-graphs "$gnum" "$svgtemp"
+elif test "$SSCONV_TEST_USE" = xlsx; then
+	ssconvert --export-graphs "$xlsx" "$svgtemp"
+elif test "$SSCONV_TEST_USE" = ods; then
+	ssconvert --export-graphs "$ods" "$svgtemp"
 else
-    echo "Unknown format <<$SSCONV_TEST_USE>!"
-    exit 1
+	echo "Unknown format <<$SSCONV_TEST_USE>!"
+	exit 1
 fi
 
 svg="$base.svg"
 mv -f "$svgtemp0" "$svg"
-# eog "$svg"
-inkscape "$svg"
+viewer="${SSCONV_TEST_VIEWER:-inkscape}"
+$viewer "$svg"
